requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
dnspython==2.6.1
idna==3.10
mysql-connector-python==9.0.0
pymongo==4.10.1
PyMySQL==1.1.1
requests==2.32.3
urllib3==2.2.3
pyspark==3.5.0